Extract parameterized inline/val tests
Eventually we'll remove inline tests but this is good for now until it's formally deprecated/removed
