+++ kotlinpoet/src/test/java/com/squareup/kotlinpoet/TypeSpecTest.kt
@@ -21,7 +21,6 @@ import com.google.testing.compile.CompilationRule
 import com.squareup.kotlinpoet.KModifier.ABSTRACT
 import com.squareup.kotlinpoet.KModifier.DATA
 import com.squareup.kotlinpoet.KModifier.IN
-import com.squareup.kotlinpoet.KModifier.INLINE
 import com.squareup.kotlinpoet.KModifier.INNER
 import com.squareup.kotlinpoet.KModifier.INTERNAL
 import com.squareup.kotlinpoet.KModifier.PRIVATE
@@ -2525,216 +2524,6 @@ class TypeSpecTest {
         |      |${"\"\"\""}.trimMargin()
         |}
         |""".trimMargin()
-    )
-  }
-
-  @Test fun validInlineClass() {
-    val guacamole = TypeSpec.classBuilder("Guacamole")
-      .primaryConstructor(
-        FunSpec.constructorBuilder()
-          .addParameter("avacado", String::class)
-          .build()
-      )
-      .addProperty(
-        PropertySpec.builder("avacado", String::class)
-          .initializer("avacado")
-          .build()
-      )
-      .addModifiers(INLINE)
-      .build()
-
-    assertThat(guacamole.toString()).isEqualTo(
-      """
-      |public inline class Guacamole(
-      |  public val avacado: kotlin.String
-      |)
-      |""".trimMargin()
-    )
-  }
-
-  @Test fun inlineClassWithInitBlock() {
-    val guacamole = TypeSpec.classBuilder("Guacamole")
-      .primaryConstructor(
-        FunSpec.constructorBuilder()
-          .addParameter("avacado", String::class)
-          .build()
-      )
-      .addProperty(
-        PropertySpec.builder("avacado", String::class)
-          .initializer("avacado")
-          .build()
-      )
-      .addInitializerBlock(CodeBlock.EMPTY)
-      .addModifiers(INLINE)
-      .build()
-
-    assertThat(guacamole.toString()).isEqualTo(
-      """
-      |public inline class Guacamole(
-      |  public val avacado: kotlin.String
-      |) {
-      |  init {
-      |  }
-      |}
-      |""".trimMargin()
-    )
-  }
-
-  class InlineSuperClass
-
-  @Test fun inlineClassWithSuperClass() {
-    assertThrows<IllegalStateException> {
-      TypeSpec.classBuilder("Guacamole")
-        .primaryConstructor(
-          FunSpec.constructorBuilder()
-            .addParameter("avocado", String::class)
-            .build()
-        )
-        .addProperty(
-          PropertySpec.builder("avocado", String::class)
-            .initializer("avocado")
-            .build()
-        )
-        .superclass(InlineSuperClass::class)
-        .addModifiers(INLINE)
-        .build()
-    }.hasMessageThat().isEqualTo("Inline classes cannot have super classes")
-  }
-
-  interface InlineSuperInterface
-
-  @Test fun inlineClassInheritsFromInterface() {
-    val guacamole = TypeSpec.classBuilder("Guacamole")
-      .primaryConstructor(
-        FunSpec.constructorBuilder()
-          .addParameter("avocado", String::class)
-          .build()
-      )
-      .addProperty(
-        PropertySpec.builder("avocado", String::class)
-          .initializer("avocado")
-          .build()
-      )
-      .addSuperinterface(InlineSuperInterface::class)
-      .addModifiers(INLINE)
-      .build()
-
-    assertThat(guacamole.toString()).isEqualTo(
-      """
-      |public inline class Guacamole(
-      |  public val avocado: kotlin.String
-      |) : com.squareup.kotlinpoet.TypeSpecTest.InlineSuperInterface
-      |""".trimMargin()
-    )
-  }
-
-  @Test fun inlineClassWithoutBackingProperty() {
-    assertThrows<IllegalArgumentException> {
-      TypeSpec.classBuilder("Guacamole")
-        .primaryConstructor(
-          FunSpec.constructorBuilder()
-            .addParameter("avocado", String::class)
-            .build()
-        )
-        .addProperty("garlic", String::class)
-        .addModifiers(INLINE)
-        .build()
-    }.hasMessageThat().isEqualTo("Inline classes must have a single read-only (val) property parameter.")
-  }
-
-  @Test fun inlineClassWithoutProperties() {
-    assertThrows<IllegalStateException> {
-      TypeSpec.classBuilder("Guacamole")
-        .primaryConstructor(
-          FunSpec.constructorBuilder()
-            .addParameter("avocado", String::class)
-            .build()
-        )
-        .addModifiers(INLINE)
-        .build()
-    }.hasMessageThat().isEqualTo("Inline classes must have at least 1 property")
-  }
-
-  @Test fun inlineClassWithMutableProperties() {
-    assertThrows<IllegalStateException> {
-      TypeSpec.classBuilder("Guacamole")
-        .primaryConstructor(
-          FunSpec.constructorBuilder()
-            .addParameter("avocado", String::class)
-            .build()
-        )
-        .addProperty(
-          PropertySpec.builder("avocado", String::class)
-            .initializer("avocado")
-            .mutable()
-            .build()
-        )
-        .addModifiers(INLINE)
-        .build()
-    }.hasMessageThat().isEqualTo("Inline classes must have a single read-only (val) property parameter.")
-  }
-
-  @Test fun inlineClassWithPrivateConstructor() {
-    val guacamole = TypeSpec.classBuilder("Guacamole")
-      .primaryConstructor(
-        FunSpec.constructorBuilder()
-          .addParameter("avocado", String::class)
-          .addModifiers(PRIVATE)
-          .build()
-      )
-      .addProperty(
-        PropertySpec.builder("avocado", String::class)
-          .initializer("avocado")
-          .build()
-      )
-      .addModifiers(INLINE)
-      .build()
-
-    assertThat(guacamole.toString()).isEqualTo(
-      """
-      |public inline class Guacamole private constructor(
-      |  public val avocado: kotlin.String
-      |)
-      |""".trimMargin()
-    )
-  }
-
-  @Test fun inlineEnumClass() {
-    val guacamole = TypeSpec.enumBuilder("Foo")
-      .primaryConstructor(
-        FunSpec.constructorBuilder()
-          .addParameter("x", Int::class)
-          .build()
-      )
-      .addEnumConstant(
-        "A",
-        TypeSpec.anonymousClassBuilder()
-          .addSuperclassConstructorParameter("%L", 1)
-          .build()
-      )
-      .addEnumConstant(
-        "B",
-        TypeSpec.anonymousClassBuilder()
-          .addSuperclassConstructorParameter("%L", 2)
-          .build()
-      )
-      .addProperty(
-        PropertySpec.builder("x", Int::class)
-          .initializer("x")
-          .build()
-      )
-      .addModifiers(INLINE)
-      .build()
-    assertThat(guacamole.toString()).isEqualTo(
-      """
-      |public enum inline class Foo(
-      |  public val x: kotlin.Int
-      |) {
-      |  A(1),
-      |  B(2),
-      |  ;
-      |}
-      |""".trimMargin()
     )
   }
