+++ kotlinpoet/src/test/java/com/squareup/kotlinpoet/ValTypeSpecTest.kt
@@ -1,0 +1,236 @@
+package com.squareup.kotlinpoet
+
+import com.google.common.truth.Truth.assertThat
+import com.squareup.kotlinpoet.KModifier.INLINE
+import com.squareup.kotlinpoet.KModifier.PRIVATE
+import org.junit.Test
+import org.junit.runner.RunWith
+import org.junit.runners.Parameterized
+
+@RunWith(Parameterized::class)
+class ValTypeSpecTest(private val useVal: Boolean) {
+
+  companion object {
+    @JvmStatic
+    @Parameterized.Parameters(name = "val={0}")
+    fun data(): Collection<Array<Any>> {
+      return listOf(
+        arrayOf(true),
+        arrayOf(false)
+      )
+    }
+  }
+
+  private val modifier = if (useVal) KModifier.VAL else INLINE
+  private val modifierString = modifier.keyword
+
+  private fun classBuilder() = if (useVal) {
+    TypeSpec.valClassBuilder("Guacamole")
+  } else {
+    TypeSpec.classBuilder("Guacamole")
+      .addModifiers(modifier)
+  }
+
+  @Test fun validInlineClass() {
+    val guacamole = classBuilder()
+      .primaryConstructor(
+        FunSpec.constructorBuilder()
+          .addParameter("avacado", String::class)
+          .build()
+      )
+      .addProperty(
+        PropertySpec.builder("avacado", String::class)
+          .initializer("avacado")
+          .build()
+      )
+      .build()
+
+    assertThat(guacamole.toString()).isEqualTo(
+      """
+      |public $modifierString class Guacamole(
+      |  public val avacado: kotlin.String
+      |)
+      |""".trimMargin()
+    )
+  }
+
+  @Test fun inlineClassWithInitBlock() {
+    val guacamole = classBuilder()
+      .primaryConstructor(
+        FunSpec.constructorBuilder()
+          .addParameter("avacado", String::class)
+          .build()
+      )
+      .addProperty(
+        PropertySpec.builder("avacado", String::class)
+          .initializer("avacado")
+          .build()
+      )
+      .addInitializerBlock(CodeBlock.EMPTY)
+      .build()
+
+    assertThat(guacamole.toString()).isEqualTo(
+      """
+      |public $modifierString class Guacamole(
+      |  public val avacado: kotlin.String
+      |) {
+      |  init {
+      |  }
+      |}
+      |""".trimMargin()
+    )
+  }
+
+  class InlineSuperClass
+
+  @Test fun inlineClassWithSuperClass() {
+    assertThrows<IllegalStateException> {
+      classBuilder()
+        .primaryConstructor(
+          FunSpec.constructorBuilder()
+            .addParameter("avocado", String::class)
+            .build()
+        )
+        .addProperty(
+          PropertySpec.builder("avocado", String::class)
+            .initializer("avocado")
+            .build()
+        )
+        .superclass(InlineSuperClass::class)
+        .build()
+    }.hasMessageThat().isEqualTo("val/inline classes cannot have super classes")
+  }
+
+  interface InlineSuperInterface
+
+  @Test fun inlineClassInheritsFromInterface() {
+    val guacamole = classBuilder()
+      .primaryConstructor(
+        FunSpec.constructorBuilder()
+          .addParameter("avocado", String::class)
+          .build()
+      )
+      .addProperty(
+        PropertySpec.builder("avocado", String::class)
+          .initializer("avocado")
+          .build()
+      )
+      .addSuperinterface(InlineSuperInterface::class)
+      .build()
+
+    assertThat(guacamole.toString()).isEqualTo(
+      """
+      |public $modifierString class Guacamole(
+      |  public val avocado: kotlin.String
+      |) : com.squareup.kotlinpoet.ValTypeSpecTest.InlineSuperInterface
+      |""".trimMargin()
+    )
+  }
+
+  @Test fun inlineClassWithoutBackingProperty() {
+    assertThrows<IllegalArgumentException> {
+      classBuilder()
+        .primaryConstructor(
+          FunSpec.constructorBuilder()
+            .addParameter("avocado", String::class)
+            .build()
+        )
+        .addProperty("garlic", String::class)
+        .build()
+    }.hasMessageThat().isEqualTo("val/inline classes must have a single read-only (val) property parameter.")
+  }
+
+  @Test fun inlineClassWithoutProperties() {
+    assertThrows<IllegalStateException> {
+      classBuilder()
+        .primaryConstructor(
+          FunSpec.constructorBuilder()
+            .addParameter("avocado", String::class)
+            .build()
+        )
+        .build()
+    }.hasMessageThat().isEqualTo("val/inline classes must have at least 1 property")
+  }
+
+  @Test fun inlineClassWithMutableProperties() {
+    assertThrows<IllegalStateException> {
+      classBuilder()
+        .primaryConstructor(
+          FunSpec.constructorBuilder()
+            .addParameter("avocado", String::class)
+            .build()
+        )
+        .addProperty(
+          PropertySpec.builder("avocado", String::class)
+            .initializer("avocado")
+            .mutable()
+            .build()
+        )
+        .build()
+    }.hasMessageThat().isEqualTo("val/inline classes must have a single read-only (val) property parameter.")
+  }
+
+  @Test
+  fun inlineClassWithPrivateConstructor() {
+    val guacamole = classBuilder()
+      .primaryConstructor(
+        FunSpec.constructorBuilder()
+          .addParameter("avocado", String::class)
+          .addModifiers(PRIVATE)
+          .build()
+      )
+      .addProperty(
+        PropertySpec.builder("avocado", String::class)
+          .initializer("avocado")
+          .build()
+      )
+      .build()
+
+    assertThat(guacamole.toString()).isEqualTo(
+      """
+      |public $modifierString class Guacamole private constructor(
+      |  public val avocado: kotlin.String
+      |)
+      |""".trimMargin()
+    )
+  }
+
+  @Test fun inlineEnumClass() {
+    val guacamole = TypeSpec.enumBuilder("Foo")
+      .addModifiers(modifier)
+      .primaryConstructor(
+        FunSpec.constructorBuilder()
+          .addParameter("x", Int::class)
+          .build()
+      )
+      .addEnumConstant(
+        "A",
+        TypeSpec.anonymousClassBuilder()
+          .addSuperclassConstructorParameter("%L", 1)
+          .build()
+      )
+      .addEnumConstant(
+        "B",
+        TypeSpec.anonymousClassBuilder()
+          .addSuperclassConstructorParameter("%L", 2)
+          .build()
+      )
+      .addProperty(
+        PropertySpec.builder("x", Int::class)
+          .initializer("x")
+          .build()
+      )
+      .build()
+    assertThat(guacamole.toString()).isEqualTo(
+      """
+      |public enum $modifierString class Foo(
+      |  public val x: kotlin.Int
+      |) {
+      |  A(1),
+      |  B(2),
+      |  ;
+      |}
+      |""".trimMargin()
+    )
+  }
+}
